refactor: create use-location-permission hook
